Event tiers: split plugin event capabilities into data / redirect / layout
GeneralEvent is now data-only (subject); redirect moves to new RedirectEvent;
LayoutEvent (redirect+layout) reparents onto it. onOrderAfterPlace & the other
place/save/calculate hooks become plain (no silent-dropped redirect); onPaymentResponse
becomes redirect-only (its layout was never rendered by PaymentController). Plugins
redirect to the gateway from onOrderProcessView, gateway-return from onPaymentResponse.
